# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SlideBanai API",
    "description": "AI-Powered Presentation Platform with Analytics & Collaboration",
    "version": "1.0.0",
    "contact": {
      "name": "SlideBanai Support",
      "url": "https://slidebanai.com/support",
      "email": "support@slidebanai.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.slidebanai.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
    "schemas": {
...
      "Template": {
        "type": "object",
        "required": ["id", "name", "category"],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Business Pitch"
          },
          "description": {
            "type": "string",
            "example": "Professional business pitch template"
          },
          "category": {
            "type": "string",
            "example": "business"
          },
          "thumbnail_url": {
            "type": "string",
            "format": "uri"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/templates": {
      "get": {
        "tags": ["Templates"],
        "summary": "Get all templates",
        "operationId": "listTemplates",
        "responses": {
          "200": {
            "description": "List of templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Template"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Templates"],
        "summary": "Create template",
        "operationId": "createTemplate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Template"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Template created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Template"
                }
              }
            }
          }
        }
      }
    },
    "/presentations/{presentationId}/templates": {
      "post": {
        "tags": ["Templates"],
        "summary": "Apply template to presentation",
        "operationId": "applyTemplate",
        "parameters": [
          {
            "name": "presentationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "template_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Template applied"
          }
        }
      }
    },
...
  }
}